refactor(llms): split src/llm.py into src/llms package
Move 818-line src/llm.py into a 7-file src/llms/ package so each
responsibility lives in its own module. Public API names are preserved
via src/llms/__init__.py. All callers updated; 161 tests still pass.
