+++ src/llms/__init__.py
@@ -1,0 +1,16 @@
+"""LLM provider package.
+
+Public re-exports. Importing submodule internals directly is fine for
+provider-specific wiring (e.g. `from src.llms.bedrock import BedrockProvider`
+in tests), but callers should prefer the names exported here.
+"""
+from src.llms.base import LLMProvider, LLMResult, ToolCall, ToolSpec
+from src.llms.factory import get_llm
+
+__all__ = [
+    "LLMProvider",
+    "LLMResult",
+    "ToolCall",
+    "ToolSpec",
+    "get_llm",
+]
